Add support for currency deletion and enhance AddDeviceFeature with SharedReader
